requirements-dev.txt
# Development environment dependencies for testing
-r requirements.txt

# Direct Dev dependencies
pytest==7.1.2
pytest-cov==3.0.0
requests-mock==1.9.3

# Secondary Dev dependencies
attrs==21.4.0
coverage==6.4.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
six==1.16.0
tomli==2.0.1
